Added error handling

+++ backend/routes/posts.js
@@ -36,7 +36,8 @@ router.post(
   const post = new Post({
     title: req.body.title,
     content: req.body.content,
-    imagePath: url + "/images/" + req.file.filename
+    imagePath: url + "/images/" + req.file.filename,
+    creator: req.userData.userId
   });
   post.save().then(createdPost => {
     res.status(201).json({
@@ -48,6 +49,8 @@ router.post(
         imagePath: createdPost.imagePath
       }
     });
+  }).catch(error => {
+    res.status(500).json({message: 'Something went wrong while creating that post.'})
   });
 });
 
@@ -76,6 +79,8 @@ router.get("",(req,res,next)=>{
       posts: fetchedPosts,
       maxPosts: count
     });
+  }).catch(error => {
+    res.status(500).json({message: "I wasn't able to get your posts :("});
   });
 });
 
@@ -87,7 +92,9 @@ router.get("/:id",(req,res,next) => {
     } else {
       res.status(404).json({message:'Post not found!'});
     }
-  })
+  }).catch(error => {
+      res.status(500).json({message: "I wasn't able to get your post :("});
+  });
 });
 
 // Update a post by id in the database
@@ -105,12 +112,18 @@ router.put(
       _id: req.body.id,
       title: req.body.title,
       content: req.body.content,
-      imagePath: imagePath
+      imagePath: imagePath,
+      creator: req.userData.userId
     });
-    Post.updateOne({_id: req.params.id}, post)
+    Post.updateOne({_id: req.params.id , creator: req.userData.userId }, post)
     .then(result => {
-      console.log(result);
-      res.status(200).json({message:'Update successful'});
+      if (result.n > 0){
+        res.status(200).json({message:'Update successful'});
+      } else {
+        res.status(404).json({message:'Not authorized!'});
+      }
+  }).catch(error => {
+    res.status(500).json({message:"Couldn't update your post :("});
   });
 });
 
@@ -119,10 +132,16 @@ router.delete(
   "/:id",
   checkAuth,
   (req,res,next) => {
-  Post.deleteOne({_id: req.params.id}).then(result => {
-    console.log(result);
-    res.status(200).json({message: "Post deleted!"});
-  });
-});
+  Post.deleteOne({_id: req.params.id , creator: req.userData.userId }).then(result => {
+    if (result.n > 0){
+      res.status(200).json({message:'Post deleted successfully'});
+    } else {
+      res.status(404).json({message:'Not authorized!'});
+    }
+  }).catch(error => {
+    res.status(500).json({message: "I wasn't able to delete this one :("});
+    });
+  }
+);
 
 module.exports = router;
